On branch janusz_ui  Changes to be committed: 	modified:   src/pdm/CLI/pdm-adduser.py 	renamed:    src/pdm/userservicedesk/app/human_resources.py -> src/pdm/userservicedesk/HRService.py 	deleted:    src/pdm/userservicedesk/app/__init__.py 	deleted:    src/pdm/userservicedesk/app/models.py 	new file:   src/pdm/userservicedesk/models.py 	deleted:    src/pdm/userservicedesk/run.py
 Untracked files:
	etc/users.auth
	etc/users.conf

The new users servive HRService.py added
