+++ src/pdm/userservicedesk/HRService.py
@@ -1,16 +1,28 @@
-""" user administration interface """
+#!/usr/bin/env python
 __author__ = 'martynia'
 
-from pdm.userservicedesk.app.models import User
+import flask
+from flask import request, abort
+from pdm.framework.FlaskWrapper import export, export_ext, startup, db_model, jsonify
+from pdm.framework.Database import from_json
+import pdm.userservicedesk.models
 
-from flask import request, jsonify, abort
 
-def manage(app):
+@export_ext("/users/api/v1.0")
+@db_model(pdm.userservicedesk.models.UserModel)
+class HRService(object):
 
-    @app.route('/users/api/v1.0/users', methods=['GET'])
+    @staticmethod
+    @export
+    def hello():
+        return jsonify('User Service Desk at your service !\n')
+
+    @staticmethod
+    @export_ext("users")
     def get_users():
         """ Get all registered users """
         # GET
+        User = request.db.tables.User
         users = User.get_all()
         results = []
 
@@ -21,7 +33,7 @@ def manage(app):
                 'name': user.name,
                 'surname' : user.surname,
                 'state' : user.state,
-                'dn' : user.dn,
+                #'dn' : user.dn,
                 'email' : user.email,
                 'password' : user.password,
                 'date_created': user.date_created,
@@ -32,28 +44,28 @@ def manage(app):
         response.status_code = 200
         return response
 
-    @app.route('/users/api/v1.0/users/<string:username>', methods=['GET'])
+    @staticmethod
+    @export_ext("users/<string:username>")
     def get_user(username):
         """
         Get user by username.
         :param username: username
         :return: json response with user data or 404 if the user does not exist
         """
+
+        User = request.db.tables.User
         user = User.query.filter_by(username=username).first()
         if not user:
             # Raise an HTTPException with a 404 not found status code
             abort(404)
-#########
-        #{k:v for k, v in user.__dict__.iteritems() if k in user.__table__.columns}
 
-#########
         result = [{
             'id': user.id,
             'username' : user.username,
             'name': user.name,
             'surname' : user.surname,
             'state' : user.state,
-            'dn' : user.dn,
+            #'dn' : user.dn,
             'email' : user.email,
             'password' : user.password,
             'date_created': user.date_created,
@@ -63,15 +75,17 @@ def manage(app):
         response.status_code = 200
         return response
 
-    @app.route('/users/api/v1.0/users', methods=['POST'])
+    @staticmethod
+    @export_ext("users", ["POST"])
     def add_user():
         """
         Add a new user.
         :return: json document with added user data.
         """
         print request.json
+        User = request.db.tables.User
         user = User(username = request.json['username'], name =request.json['name'], surname = request.json['surname'],
-                    email = request.json["email"], state = request.json['state'], dn = request.json['dn'], password = request.json['password'])
+                    email = request.json["email"], state = 0, password = request.json['password'])
         user.save()
         response = jsonify([{
             'id': user.id,
@@ -79,7 +93,7 @@ def manage(app):
             'username': user.username,
             'surname': user.surname,
             'state' : user.state,
-            'dn' : user.dn,
+            #'dn' : user.dn,
             'email' : user.email,
             'password' :user.password,
             'date_created': user.date_created,
@@ -88,13 +102,17 @@ def manage(app):
         response.status_code = 201
         return response
 
-    @app.route('/users/api/v1.0/users/<string:username>', methods=['PUT'])
+
+    @staticmethod
+    @export_ext("users/<string:username>", ["PUT"])
     def update_user(username):
         """
         Update an existing user.
         :param username: username of the user to be updated
         :return: json doc of the updated user or 404 if the user does not exist
         """
+
+        User = request.db.tables.User
         user = User.query.filter_by(username=username).first()
         if not user:
             # Raise an HTTPException with a 404 not found status code
@@ -113,7 +131,7 @@ def manage(app):
             'username': user.username,
             'surname': user.surname,
             'state' : user.state,
-            'dn' : user.dn,
+            #'dn' : user.dn,
             'email' : user.email,
             'password' :user.password,
             'date_created': user.date_created,
@@ -122,14 +140,17 @@ def manage(app):
         response.status_code = 200
         return response
 
-    @app.route('/users/api/v1.0/users/<string:username>', methods=['DELETE'])
+    @staticmethod
+    @export_ext("users/<string:username>", ["PUT"])
     def delete_user(username):
         """
         Delete a user
         :param username: a username of an existing user
         :return: code 200 if successful, 404 if the user does not exist
         """
+        User = request.db.tables.User
         user = User.query.filter_by(username=username).first()
+
         if not user:
             # Raise an HTTPException with a 404 not found status code
             abort(404)
@@ -141,5 +162,3 @@ def manage(app):
 
         response.status_code = 200
         return response
-
-    return app
